+++ aom/cmd/ls.go
@@ -1,37 +1,32 @@
-// Copyright © 2018 NAME HERE <EMAIL ADDRESS>
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
 package cmd
 
 import (
 	"fmt"
 
 	"github.com/spf13/cobra"
+
+	"github.com/philippgille/apiomat-go/aomc"
+	"github.com/philippgille/apiomat-go/aoms"
 )
 
 // lsCmd represents the ls command
 var lsCmd = &cobra.Command{
 	Use:   "ls",
-	Short: "A brief description of your command",
-	Long: `A longer description that spans multiple lines and likely contains examples
-and usage of using your command. For example:
+	Short: "Lists classes of a given module",
+	Long: `
+ls is a command for listing classes of a given module.
 
-Cobra is a CLI library for Go that empowers applications.
-This application is a tool to generate the needed files
-to quickly create a Cobra application.`,
+For example, to list all classes of module "MyModule":
+aom class ls --module "MyModule" --baseUrl "https://apiomat.yourcompany.com/yambas/rest" --username "john" --password "secret"`,
 	Run: func(cmd *cobra.Command, args []string) {
-		fmt.Println("ls called")
+		client := aomc.NewDefaultClient(baseUrl, username, password, aoms.System(system))
+
+		classes, err := client.GetClasses(module)
+		if err != nil {
+			panic(err)
+		}
+
+		fmt.Printf("Classes of module %v: %+v", module, classes)
 	},
 }